feat(infra): add S2S validation client and initial session repository
- Implement a generic HTTP client in 'packages/utils' for server-to-server communication. This client will be used for validating player access with an external authentication service.
- Add an initial GORM-based repository in 'game_session/infrastructure/trash' to handle access checks and retrieve session host information from a PostgreSQL database.
- Include self-signed SSL certificate and key ('cert.pem', 'key.pem') in 'tmp/cert' for enabling local HTTPS/WSS development and testing.

The new S2S API client provides a standardized way to perform authenticated requests to other backend services. The 'trash' repository establishes the pattern for interacting with the main backend's database to verify tickets and session ownership. SSL certificates are added to facilitate a secure local development environment that mirrors production conditions.

+++ src/modules/hub/games/sales_courage/dto/dto.go
@@ -1,6 +1,6 @@
 package sales_courage_dto
 
 type AddCoins struct {
-	PlayerID string  `json:"player_id"`
-	Coins    float64 `json:"coins"`
+	PlayerID string `json:"player_id"`
+	Coins    int    `json:"coins"`
 }
